Transfer Learning/requirements.txt
torch
torchvision
matplotlib
numpy
Pillow
scikit-learn